Enhance product details with accordion & table
This update introduces two major UI improvements to the product details page, aimed at enhancing user experience by presenting information in a more organized and accessible manner. The addition of an accordion component allows for a cleaner presentation of content like materials, which users can now expand or collapse as needed. The implementation of a table for displaying product dimensions provides a structured and easy-to-read format, significantly improving the readability of critical product specifications.

Additionally, this commit includes a dependency update, adding `@radix-ui/react-accordion` to support the new accordion component. The Tailwind configuration has been extended with new keyframes and animations for smooth accordion transitions, elevating the visual experience. Furthermore, a minor but crucial UI fix in the navigation bar now wraps the logo in an anchor tag, ensuring it correctly serves as a home link.

This cohesive set of improvements not only elevates the product page's functionality but also its aesthetic appeal, contributing to a more engaging and informative user shopping experience.